Start properly organizing this code (meeting api)

+++ src/controllers/MeetingController.ts
@@ -4,7 +4,16 @@ import db from "../sequelize/models";
 import { handleException } from "./utils/ErrorHandler";
 import { Meeting } from "../models/Meeting";
 
-export async function getMeetings(req: Request, res: Response) {
+const { param, body } = require('express-validator');
+export const idValidator = [
+    param('id').exists().isInt().withMessage("Invalid meeting id")
+]
+
+export const meetingValidator = [
+    // TODO
+]
+
+export async function getAllMeetings(req: Request, res: Response) {
     const { agentid } = req.query;
     if (agentid) {
         getMeetingsOfAgent(req, res);
@@ -14,7 +23,7 @@ export async function getMeetings(req: Request, res: Response) {
     const meetingRepository = new MeetingRepository(db);
     try {
         const meetings = await meetingRepository.getAll();
-        res.status(200).json(meetings);
+        res.status(200).json(meetings); // TODO USE DTO
     }
     catch (error) {
         handleException(res, error);
@@ -27,7 +36,25 @@ async function getMeetingsOfAgent(req: Request, res: Response) {
 
     try {
         const meetings = await meetingRepository.getMeetingsOfAgent(agentid as string);
-        res.status(200).json(meetings);
+        res.status(200).json(meetings); // TODO USE DTO
+    }
+    catch (error) {
+        handleException(res, error);
+    }
+}
+
+export async function getMeetingById(req: Request, res: Response) {
+    const meetingRepository = new MeetingRepository(db);
+    const { id } = req.params;
+    const idInt = parseInt(id);
+
+    try {
+        const meeting = await meetingRepository.findMeetingById(idInt);
+        if (!meeting) {
+            res.status(404).json({ message: "Meeting not found" });
+            return;
+        }
+        res.status(200).json(meeting); // TODO USE DTO
     }
     catch (error) {
         handleException(res, error);
@@ -41,23 +68,49 @@ export async function addMeeting(req: Request, res: Response) {
     try {
         let meeting = new Meeting(time, place, client, phoneOperator, salesAgent, worker);
         meeting = await meetingRepository.save(meeting);
-        res.status(201).json(meeting);
+        res.status(201).json(meeting); // TODO USE DTO
     }
     catch (error) {
         handleException(res, error);
     }
+}
 
+export async function updateMeeting(req: Request, res: Response) {
+    const meetingRepository = new MeetingRepository(db);
+    const { meeting_id, time, place, client, phoneOperator, salesAgent, worker } = req.body;
+
+    const idInt = parseInt(meeting_id);
+
+    try {
+        let meeting = await meetingRepository.findMeetingById(idInt);
+        if (!meeting) {
+            res.status(404).json({ message: "Meeting not found" });
+            return;
+        }
+
+        meeting = new Meeting(time, place, client, phoneOperator, salesAgent, worker);
+        await meetingRepository.save(meeting);
+        res.status(200).json(meeting); // TODO USE DTO
+    }
+    catch (error) {
+        handleException(res, error);
+    }
 }
 
 export async function deleteMeeting(req: Request, res: Response) {
     const meetingRepository = new MeetingRepository(db);
-    const { meeting_id } = req.body;
+    const { id } = req.params;
 
-    const idInt = parseInt(meeting_id as string);
-
+    const idInt = parseInt(id);
     try {
-        const result = await meetingRepository.delete(idInt);
-        res.status(204).json();
+        let meeting = await meetingRepository.findMeetingById(idInt);
+        if (!meeting) {
+            res.status(404).json({ message: "Meeting not found" });
+            return;
+        }
+
+        await meetingRepository.delete(idInt);
+        res.status(204).send();
     }
     catch(error) {
         handleException(res, error);
